refactor: redis 관련 설정 통합
