+++ muckpot-infra/src/main/kotlin/com/yapp/muckpot/redis/RedisConfig.kt
@@ -1,5 +1,8 @@
 package com.yapp.muckpot.redis
 
+import org.redisson.Redisson
+import org.redisson.api.RedissonClient
+import org.redisson.config.Config
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -20,6 +23,7 @@ class RedisConfig(
     fun redisConnectionFactory(): RedisConnectionFactory {
         return LettuceConnectionFactory(redisHost, redisPort)
     }
+    private val REDISSON_HOST_PREFIX = "redis://"
 
     @Bean
     fun redisTemplate(): RedisTemplate<String, Any> {
@@ -27,5 +31,12 @@ class RedisConfig(
         redisTemplate.setConnectionFactory(redisConnectionFactory())
         redisTemplate.keySerializer = StringRedisSerializer()
         return redisTemplate
+    }
+
+    @Bean
+    fun redissonClient(): RedissonClient {
+        val config = Config()
+        config.useSingleServer().address = "$REDISSON_HOST_PREFIX$redisHost:$redisPort"
+        return Redisson.create(config)
     }
 }
