--- muckpot-infra/src/main/kotlin/com/yapp/muckpot/redis/RedissonConfig.kt
@@ -1,26 +1,0 @@
-package com.yapp.muckpot.redis
-
-import org.redisson.Redisson
-import org.redisson.api.RedissonClient
-import org.redisson.config.Config
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class RedissonConfig(
-    @Value("\${spring.redis.host}")
-    private val redisHost: String,
-
-    @Value("\${spring.redis.port}")
-    private val redisPort: Int
-) {
-    private val REDISSON_HOST_PREFIX = "redis://"
-
-    @Bean
-    fun redissonClient(): RedissonClient {
-        val config = Config()
-        config.useSingleServer().address = "$REDISSON_HOST_PREFIX$redisHost:$redisPort"
-        return Redisson.create(config)
-    }
-}
